requirements.txt
black==23.1.0
Flask==2.2.3
isort==5.12.0
jsonschema==4.17.3
mpmath==1.2.1
mypy-extensions==1.0.0
numpy==1.24.2
pandas==1.5.3
pylint==2.16.2
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
scipy==1.10.0
six==1.16.0